Sentence-level edits
I tightened some of the languages. 

ALSO - I was intending to delete the references slide (14) entirely and have already distributed that content to slides 12 and 13, where it seemed more appropriate. I did not delete the "References" slide, however, because I see this kind of code in this file: 

!---

## Basis Risk: Prices by Futures Maturity

```{r, echo=FALSE, message=FALSE, warning=FALSE, cache = FALSE}
#{{{

etc. 

Wanting to avoid messing things up, I left the "References" slide for the moment, BUT it needs to be deleted.